refactor(react): extract the Mux storyboard track into its own module
`MuxStoryboard` was a private helper inside the hls.js-backed `MuxVideo`, but
nothing about it is engine-specific — it reads Mux content data and the stream
type. Moving it up beside the flavors lets a second one render the same track
without duplicating the deferred-notification subscription.

Its media prop is now a structural contract rather than the concrete
`MuxMedia`, so the component carries no engine, and the live check reads
`MediaStreamTypes` instead of the hls.js re-export.
